Added Cinemachine Package + Camera, updated player object
Added Cinemachine package to create a camera component to follow and look at the player's vehicle.

Created the camera and specified orbital radii for top, mid, and bottom riggings.

Updated player character model to a vehicle (from a cube), including wheel colliders and meshes.

Player movement updated to use transformations, but must be updated to use Unity's car physics in the future.